refactor: Move Deposit Function to core
