+++ core/ui/vault/deposit/DepositConfirmButton/config.ts
