+++ core/ui/vault/deposit/DepositForm/ActionSpecific/StakeSpecific/StakeSpecific.tsx
