+++ core/ui/vault/deposit/DepositForm/ActionSpecific/StakeSpecific/StakeTokenExplorer.tsx
