+++ core/ui/vault/deposit/DepositVerify/DepositVerify.styled.tsx
