+++ core/ui/vault/deposit/DepositVerify/utils.ts
