+++ core/ui/vault/deposit/fee/DepositFeeValue.tsx
