+++ core/ui/vault/deposit/fee/DepositFiatFeeValue.tsx
